Move staking test
